fix: show search in description (#56)
* fix: show search in description

When searching for a post, the description was not changing.
The static description was being used and it sometimes doesn't had the
searched term that is in the `full` property. So, now in the search
route, the description is set to include the searched term, and the ellipsis.

* refactor: do not include title in words list

* refactor: validate if search term in present outside of title

* fix: add includeSearch function to check for the search term

Accepting reviewer suggestion.

Co-authored-by: Nícolas Gabriel <nickygamer023@gmail.com>

* fix: prettier problems

* fix: use includeSearch to find search term index

* refactor: more words in beginning if text is too short

Also avoid a negative start.

* fix: ellipsis in the beginning not working as expected

Added a verification for the first word, so if it's an emoji, a new line char or a whitespace, it will not add the ellipsis. Also if index is 0, it will not add the ellipsis.

---------

Co-authored-by: Nícolas Gabriel <nickygamer023@gmail.com>